hashCode + coefficient return type fix
implemented abstract hashCode method. Changed coefficient return type from int to double to reflect rep capabilities.

+++ src/MathFunctionKernel.java
@@ -38,7 +38,7 @@ public interface MathFunctionKernel extends Standard<MathFunctionKernel> {
      * @ensures coefficient of x = c where degree n = cx^n
      *
      */
-    int coefficient(int degree);
+    double coefficient(int degree);
 
     /**
      * Retirns size of the poloynomial.
